[fix] change used button result according to user's role
